# Conflicts:
#	src/main/java/com/example/TucShopBackend/Models/Category.java
#	src/main/resources/application-dev.properties

+++ src/main/java/com/example/TucShopBackend/Services/CategoryService.java
@@ -44,8 +44,8 @@ public class CategoryService {
     String profile;
 
     //serverfile.path
-    @Value("${serverfile.path}")
-    String serverFilePath;
+//    @Value("${serverfile.path}")
+//    String serverFilePath;
 
     @Autowired
     private CloudinaryService cloudinaryService;
@@ -94,7 +94,7 @@ public class CategoryService {
     public Boolean saveCategoryImage(MultipartFile file, String name, String unique){
         try {
 
-            String UPLOADED_FOLDER_NEW = serverFilePath+"serverFiles//"+name+"//";
+            String UPLOADED_FOLDER_NEW = CustomConstants.SERVER_PATH+"//"+"serverFiles//"+name+"//";
 
             File dir = new File(UPLOADED_FOLDER_NEW);
             dir.setExecutable(true);
@@ -123,7 +123,7 @@ public class CategoryService {
     public ResponseEntity<InputStreamResource> getCategoryImage(String filename, String category) throws IOException{
 
 
-        String filepath = serverFilePath+"serverFiles//"+category+"//"+filename;
+        String filepath = CustomConstants.SERVER_PATH+"//"+"serverFiles//"+category+"//"+filename;
 
         File f = new File(filepath);
         Resource file = new UrlResource(f.toURI());
@@ -192,15 +192,23 @@ public class CategoryService {
 
     }
 
-    public  ApiResponse  deleteCategory (Long id){
-     categoryRepository.deleteById(id);
-
-    return  new ApiResponse (Status.Status_Ok, CustomConstants.CAT_DELETE, null, getAll());
-
+    public  ApiResponse  deleteCategory (Long id) {
+        Optional<Category> category = categoryRepository.findById(id);
+        boolean  check=false;
+        if (category.isPresent()) {
+            String folderPath = CustomConstants.SERVER_PATH+"//"+"serverFiles//"+category.get().getName();
+            File folder = new File(folderPath);
+            check =deleteDirectory(folder);
+            categoryRepository.deleteById(id);
+        }
+            if(check) {
+                return new ApiResponse(Status.Status_Ok, CustomConstants.CAT_DELETE, null, getAll());
+            }else {
+                return new ApiResponse(Status.Status_ERROR, CustomConstants.CATIMAGE_ERROR, null, getAll());
+            }
     }
 
-
-   public ApiResponse <Category> deleteAll (){
+    public ApiResponse <Category> deleteAll (){
     categoryRepository.deleteAll();
 
      return new ApiResponse  (Status.Status_Ok, CustomConstants.CAT_DELETE, null);
@@ -222,5 +230,17 @@ public class CategoryService {
         categoryRepository.save(category1);
         return new ApiResponse(200, CustomConstants.CAT_UPDATE, category1);
     }
+
+    private boolean deleteDirectory(File directoryToBeDeleted) {
+        File[] allContents = directoryToBeDeleted.listFiles();
+        if (allContents != null) {
+            for (File file : allContents) {
+                deleteDirectory(file);
+            }
+        }
+        return directoryToBeDeleted.delete();
+    }
+
+
 
 }
